requirements.txt
ray[default]
tensordict>=0.10.0
pyzmq
hydra-core
numpy<2.0.0
msgspec
psutil
